refactor: extract Show instances from Type.hs into separate Show.hs module
- Create new Core.Show module containing all Show instances for Term, Book, Span, Error, and Ctx
- Move helper functions (prettyStr, prettyCtr) used by Show instances
- Update imports in modules that use show on Core.Type types
- Add Core.Show to exposed-modules in cabal file
- This separation improves module organization and reduces Type.hs size
