+++ src/Core/Parse/Book.hs
@@ -22,6 +22,7 @@ import Core.Adjust
 import Core.Parse
 import Core.Parse.Term (parseTerm, parseTermBefore)
 import Core.Type
+import Core.Show
 
 -- | Book parsing
